suzdal_full_guide_refine/requirements.txt
fastapi
uvicorn
requests
pandas
python-dotenv
langchain
langchain_gigachat
tenacity
